Basic Low Fidelity Figma Designs Complete
Redesigned the figma design from yesterdays completion with better details and added the rest of the pages that will be included on the dashboard like login etc. Tried not to spend too much time on these but wanted a decent visual to guide the development of the project as it definitely made me think of some more things that i wouldve missed out (analytics page for example).