Updated Designs + Requirements
Have been thinking about different tools and decided upon swapping the postgresql db for firebase as its much more efficient and will save a lot of time and hassle, especially with concerns around security. Went back to the requirements to inlcude this and also included the analytics and settings into the designs which I decided upon yesterday from doing the low-fid.