Add log-forwarder and sample log files
Introduced log-forwarder.py for copying and log files between raw and cleaned directories. So far it scans the src dir, and copies each file (if its not a duplicate) to the "clean directory". Added sample log files for apache_attack, mixed_threats, and server_attack as well as a pseudo-code reference. Updated frontend to reposition TestConnection and simplified its styling.